Split function inference conflict tests
