+++ tests/generic_diagnostics/inference_conflicts/functions.rs
@@ -1,0 +1,6 @@
+#[path = "functions/basic.rs"]
+mod basic;
+#[path = "functions/compound_params.rs"]
+mod compound_params;
+#[path = "functions/generic_params.rs"]
+mod generic_params;
