+++ tests/generic_diagnostics/inference_conflicts/methods/compound_params.rs
@@ -1,4 +1,4 @@
-use super::super::*;
+use super::super::super::*;
 
 #[test]
 fn generic_method_inference_conflict_through_function_type_is_error() {
